+++ apps/server/src/modules/companies/application/use-cases/create-company.use-case.ts
@@ -4,15 +4,28 @@ import {
   type CompanyRepository,
   type CompanyUpsertParams,
 } from "../services/company.repository";
+import {
+  DOMAIN_EVENT_PUBLISHER,
+  type DomainEventPublisher,
+} from "../services/domain-event-publisher";
+import { CompanyAggregate } from "../../domain/aggregates/company.aggregate";
 
 @Injectable()
 export class CreateCompanyUseCase {
   public constructor(
     @Inject(COMPANY_REPOSITORY)
     private readonly companyRepository: CompanyRepository,
+    @Inject(DOMAIN_EVENT_PUBLISHER)
+    private readonly domainEventPublisher: DomainEventPublisher,
   ) {}
 
-  public execute(params: CompanyUpsertParams) {
-    return this.companyRepository.create(params);
+  public async execute(params: CompanyUpsertParams) {
+    const company = await this.companyRepository.create(params);
+
+    await this.domainEventPublisher.publishAll([
+      CompanyAggregate.fromRecord(company).createdEvent(),
+    ]);
+
+    return company;
   }
 }